meet-buddy/requirements.txt
pyannote.audio
cffi
ffmpeg-python
openai-whisper
pandas
matplotlib